usb: dwc3: gadget: Prevent core from processing stale TRBs
With CPU re-ordering on write instructions, there might
be a chance that the HWO is set before the TRB is updated
with the new mapped buffer address.
And in the case where core is processing a list of TRBs
it is possible that it fetched the TRBs when the HWO is set
but before the buffer address is updated.
Prevent this by adding a memory barrier before the HWO
is updated to ensure that the core always process the
updated TRBs.

Mot-CRs-fixed: (CR)

Change-Id: If81f2d213fb7b26384ab7c335076ec8f21e4922e
Signed-off-by: Pratham Pratap <prathampratap@codeaurora.org>
Reviewed-on: https://gerrit.mot.com/1422072
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Xiaojun Ji <jixj@motorola.com>
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key

 Conflicts:
	drivers/usb/dwc3/ep0.c

Change-Id: Ifd7e7a1654843b12f477b6212bdbc3983c89a21b