techpack: audio: dsp: use crus_afe_callback directly
When compiling built-in, this causes the symbol name to match the
symbol in the driver that is setting the callback (in this case, it is
msm-cirrus-playback).

Since both drivers are being built in, there's no reason to set a
callback from a separate function. So remove the logic and switch to
extern to use the function directly.

Change-Id: I42e098f84f01e82a04e5407e052a1a6c8b615e5a